reworked all static pages into a static pages module rather than some in app module and some in articles module
