+++ src/app/app-routing.module.ts
@@ -1,17 +1,10 @@
 import { NgModule }             from '@angular/core'
 import { RouterModule, Routes } from '@angular/router'
 import { HomeComponent } from './home/home.component'
-import { AboutComponent } from './about/about.component'
-import { ContactComponent } from './contact/contact.component'
-import { LegalComponent } from './legal/legal.component'
 
 const routes: Routes = [
-  { path: '', component: HomeComponent },
-  { path: 'about', component: AboutComponent},
-  { path: 'contact', component: ContactComponent},
-  { path: 'legal', component: LegalComponent},
-  { path: 'articles', loadChildren: './articles/articles.module#ArticlesModule' },
-  { path: '**', redirectTo: '', pathMatch: 'full'}
+  { path: '', pathMatch:'full', component: HomeComponent },
+  { path: '', loadChildren: './staticpages/staticpages.module#StaticPagesModule' },
 ];
 
 @NgModule({
