--- src/app/articles/articles.module.spec.ts
@@ -1,13 +1,0 @@
-import { ArticlesModule } from './articles.module';
-
-describe('ArticlesModule', () => {
-  let articlesModule: ArticlesModule;
-
-  beforeEach(() => {
-    articlesModule = new ArticlesModule();
-  });
-
-  it('should create an instance', () => {
-    expect(articlesModule).toBeTruthy();
-  });
-});
