--- src/app/articles/articles.module.ts
@@ -1,20 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { routes } from './articles.routing';
-import { RouterModule } from '@angular/router';
-import { DelayingSocialSecurity8ReturnComponent } from './delaying-social-security8-return/delaying-social-security8-return.component';
-import { ArticlesHomeComponent } from './articles-home/articles-home.component';
-import { SharedModule } from '../shared/shared.module';
-import { SimilarPIAsComponent } from './similar-pias/similar-pias.component';
-import { SpousalBenefitCalculationComponenentComponent } from './spousal-benefit-calculation-componenent/spousal-benefit-calculation-componenent.component';
-
-
-@NgModule({
-  imports: [
-    CommonModule,
-    RouterModule.forChild(routes),
-    SharedModule
-  ],
-  declarations: [DelayingSocialSecurity8ReturnComponent, ArticlesHomeComponent, SimilarPIAsComponent, SpousalBenefitCalculationComponenentComponent]
-})
-export class ArticlesModule { }
