+++ src/app/staticpages/about/about.component.ts
