+++ src/app/staticpages/contact/contact.component.ts
