+++ src/app/staticpages/legal/legal.component.html
