+++ src/app/staticpages/legal/legal.component.ts
