+++ src/app/staticpages/staticpages.module.spec.ts
@@ -1,0 +1,13 @@
+import { StaticPagesModule } from './staticpages.module';
+
+describe('StaticPagesModule', () => {
+  let articlesModule: StaticPagesModule;
+
+  beforeEach(() => {
+    articlesModule = new StaticPagesModule();
+  });
+
+  it('should create an instance', () => {
+    expect(articlesModule).toBeTruthy();
+  });
+});
